+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Changed
+
+- [#311][]: Changed the error message when defining the default value for a
+  hash.
+
 # [2.1.3][] (2015-10-02)
 
 ## Fixed